Refactor: migrate remaining TMR examples to tests/st and upgrade paged attention
- Move spmd_*, mixed_example from examples/tmr/ to tests/st/tmr/
- Remove duplicate HBG paged_attention from examples/ (already in tests/st/)
- Remove old TMR paged_attention from tests/st/ (kept in examples/ as evolving reference)
- Upgrade TMR paged_attention: fp16 -> bfloat16, multi-tile dispatch (16x128, 64x64),
  production-scale cases (batch=256, head_dim=128/256), tighter tolerances (1e-3)
- Add small-tile (16,16,16) dispatch path to HBG paged_attention kernels
  with SmallCase1/SmallCase2 sim-compatible test cases
