--- examples/a5/host_build_graph/paged_attention/kernels/aiv/aiv_online_update.cpp
@@ -1,230 +1,0 @@
-/*
- * Copyright (c) PyPTO Contributors.
- * This program is free software, you can redistribute it and/or modify it under the terms and conditions of
- * CANN Open Software License Agreement Version 2.0 (the "License").
- * Please refer to the License for details. You may not use this file except in compliance with the License.
- * THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND, EITHER EXPRESS OR IMPLIED,
- * INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT, MERCHANTABILITY, OR FITNESS FOR A PARTICULAR PURPOSE.
- * See LICENSE in the root of the software repository for the full text of the License.
- * -----------------------------------------------------------------------------------------------------------
- */
-// Online Softmax Update + Normalize Kernel (AIV)
-//
-// Fixed tile size: oi/oi_new are (16, 16), mij/lij/mi/li are 16-element vectors
-//
-// Scalar layout strategy:
-//   M scalar floats stored contiguously in GM can be loaded as either:
-//   - ND (kScalarRows, kScalarCols) RowMajor for element-wise ops (TMAX, TSUB, TEXP, TMUL, TADD)
-//   - DN (kAlignedRows, 1) ColMajor for row-broadcast ops (TROWEXPANDMUL, TROWEXPANDDIV)
-//   Conversion between layouts uses GM round-trip: ND TSTORE -> DN TLOAD.
-
-#include <cstdint>
-#include <pto/pto-inst.hpp>
-
-using namespace pto;
-
-#ifndef __gm__
-#define __gm__
-#endif
-
-#ifndef __aicore__
-#define __aicore__ [aicore]
-#endif
-
-static __aicore__ void online_update_impl(
-    __gm__ uint8_t *mij_raw, __gm__ uint8_t *lij_raw, __gm__ uint8_t *oi_new_raw, __gm__ uint8_t *mi_raw,
-    __gm__ uint8_t *li_raw, __gm__ uint8_t *oi_raw, int is_first, int is_last, __gm__ uint8_t *dst_raw
-) {
-    constexpr int M = 16, N = 16;
-
-    __gm__ float *mij_ptr = reinterpret_cast<__gm__ float *>(mij_raw);
-    __gm__ float *lij_ptr = reinterpret_cast<__gm__ float *>(lij_raw);
-    __gm__ float *oi_new_ptr = reinterpret_cast<__gm__ float *>(oi_new_raw);
-    __gm__ float *mi_ptr = reinterpret_cast<__gm__ float *>(mi_raw);
-    __gm__ float *li_ptr = reinterpret_cast<__gm__ float *>(li_raw);
-    __gm__ float *oi_ptr = reinterpret_cast<__gm__ float *>(oi_raw);
-    __gm__ float *dst_ptr = reinterpret_cast<__gm__ float *>(dst_raw);
-
-    // Scalar tile dimensions for RowMajor layout:
-    // kScalarCols = 32 bytes / 4 bytes per float = 8 floats per row (one 32-byte block)
-    // kScalarRows = M / 8 (M=16 -> 2 rows)
-    constexpr int kScalarCols = 32 / sizeof(float);
-    constexpr int kScalarRows = M / kScalarCols;
-    // Aligned rows for ColMajor DN tiles (32-byte alignment)
-    constexpr int kAlignedRows = ((M * sizeof(float) + 31) / 32) * (32 / sizeof(float));
-
-    // --- GlobalTensor types ---
-
-    // Data (M, N) RowMajor
-    using GlobalDataMxN = GlobalTensor<float, Shape<1, 1, 1, M, N>, pto::Stride<1, 1, 1, N, 1>>;
-
-    // Scalar ND: M contiguous floats as (kScalarRows, kScalarCols) RowMajor
-    using GlobalScalarND =
-        GlobalTensor<float, Shape<1, 1, 1, kScalarRows, kScalarCols>, pto::Stride<1, 1, 1, kScalarCols, 1>>;
-
-    // Scalar DN: same M contiguous floats as (kAlignedRows, 1) ColMajor
-    using GlobalScalarDN = GlobalTensor<float, Shape<1, 1, 1, kAlignedRows, 1>, pto::Stride<1, 1, 1, 1, 1>, Layout::DN>;
-
-    // --- GlobalTensor instances ---
-
-    GlobalDataMxN oiNewGlobal(oi_new_ptr);
-    GlobalDataMxN oiGlobal(oi_ptr);
-    GlobalDataMxN dstGlobal(dst_ptr);
-
-    // ND globals for scalar element-wise operations
-    GlobalScalarND mijGlobalND(mij_ptr);
-    GlobalScalarND lijGlobalND(lij_ptr);
-    GlobalScalarND miGlobalND(mi_ptr);
-    GlobalScalarND liGlobalND(li_ptr);
-
-    // DN globals aliased to same GM for ColMajor reload (used after ND TSTORE)
-    GlobalScalarDN mijGlobalDN(mij_ptr);
-    GlobalScalarDN lijGlobalDN(lij_ptr);
-    GlobalScalarDN liGlobalDN(li_ptr);
-
-    // --- Tile types ---
-
-    using TileDataMxN = Tile<TileType::Vec, float, M, N, BLayout::RowMajor, M, N>;
-    using TileScalarND =
-        Tile<TileType::Vec, float, kScalarRows, kScalarCols, BLayout::RowMajor, kScalarRows, kScalarCols>;
-    using TileScalarDN = Tile<TileType::Vec, float, kAlignedRows, 1, BLayout::ColMajor, M, 1>;
-
-    // --- UB memory layout ---
-
-    constexpr int kDataBytes = M * N * sizeof(float);
-    constexpr int kScalarNDBytes = kScalarRows * kScalarCols * sizeof(float);
-    constexpr int kScalarDNBytes = kAlignedRows * sizeof(float);
-
-    // Data tiles
-    TileDataMxN oiNewTile;
-    TileDataMxN oiTile;
-
-    // Scalar ND tiles for element-wise arithmetic
-    TileScalarND mijND, lijND, miND, liND;
-    TileScalarND miNewND, alphaND, betaND, tmpND;
-
-    // Scalar DN tiles for TROWEXPAND operations
-    TileScalarDN alphaDN, betaDN, liDN;
-
-    TASSIGN(oiNewTile, 0);
-    TASSIGN(oiTile, kDataBytes);
-    TASSIGN(mijND, 2 * kDataBytes);
-    TASSIGN(lijND, 2 * kDataBytes + kScalarNDBytes);
-    TASSIGN(miND, 2 * kDataBytes + 2 * kScalarNDBytes);
-    TASSIGN(liND, 2 * kDataBytes + 3 * kScalarNDBytes);
-    TASSIGN(miNewND, 2 * kDataBytes + 4 * kScalarNDBytes);
-    TASSIGN(alphaND, 2 * kDataBytes + 5 * kScalarNDBytes);
-    TASSIGN(betaND, 2 * kDataBytes + 6 * kScalarNDBytes);
-    TASSIGN(tmpND, 2 * kDataBytes + 7 * kScalarNDBytes);
-    TASSIGN(alphaDN, 2 * kDataBytes + 8 * kScalarNDBytes);
-    TASSIGN(betaDN, 2 * kDataBytes + 8 * kScalarNDBytes + kScalarDNBytes);
-    TASSIGN(liDN, 2 * kDataBytes + 8 * kScalarNDBytes + 2 * kScalarDNBytes);
-
-    if (is_first) {
-        // --- First block: copy inputs to accumulators ---
-        TLOAD(oiNewTile, oiNewGlobal);
-        TLOAD(mijND, mijGlobalND);
-        TLOAD(lijND, lijGlobalND);
-        set_flag(PIPE_MTE2, PIPE_V, EVENT_ID0);
-        wait_flag(PIPE_MTE2, PIPE_V, EVENT_ID0);
-
-        // Passthrough to MTE3 (no V compute needed)
-        set_flag(PIPE_V, PIPE_MTE3, EVENT_ID0);
-        wait_flag(PIPE_V, PIPE_MTE3, EVENT_ID0);
-        TSTORE(miGlobalND, mijND);    // mi = mij
-        TSTORE(liGlobalND, lijND);    // li = lij
-        TSTORE(oiGlobal, oiNewTile);  // oi = oi_new
-
-        if (is_last) {
-            // Single block: normalize dst = oi_new / lij
-            // lij stored to li buffer in ND format; reload as DN for TROWEXPANDDIV
-            set_flag(PIPE_MTE3, PIPE_MTE2, EVENT_ID0);
-            wait_flag(PIPE_MTE3, PIPE_MTE2, EVENT_ID0);
-            TLOAD(liDN, liGlobalDN);
-            set_flag(PIPE_MTE2, PIPE_V, EVENT_ID1);
-            wait_flag(PIPE_MTE2, PIPE_V, EVENT_ID1);
-            TROWEXPANDDIV(oiNewTile, oiNewTile, liDN);
-            set_flag(PIPE_V, PIPE_MTE3, EVENT_ID1);
-            wait_flag(PIPE_V, PIPE_MTE3, EVENT_ID1);
-            TSTORE(dstGlobal, oiNewTile);
-        }
-    } else {
-        // --- Subsequent blocks: accumulate ---
-
-        // Phase 1: Load all inputs
-        TLOAD(oiNewTile, oiNewGlobal);
-        TLOAD(oiTile, oiGlobal);
-        TLOAD(mijND, mijGlobalND);
-        TLOAD(lijND, lijGlobalND);
-        TLOAD(miND, miGlobalND);
-        TLOAD(liND, liGlobalND);
-        set_flag(PIPE_MTE2, PIPE_V, EVENT_ID0);
-        wait_flag(PIPE_MTE2, PIPE_V, EVENT_ID0);
-
-        // Phase 2: Scalar arithmetic in RowMajor (kScalarRows, kScalarCols)
-        TMAX(miNewND, miND, mijND);    // mi_new = max(mi, mij)
-        TSUB(alphaND, miND, miNewND);  // alpha = mi - mi_new
-        TEXP(alphaND, alphaND);        // alpha = exp(mi - mi_new)
-        TSUB(betaND, mijND, miNewND);  // beta = mij - mi_new
-        TEXP(betaND, betaND);          // beta = exp(mij - mi_new)
-        TMUL(liND, alphaND, liND);     // li = alpha * li
-        TMUL(tmpND, betaND, lijND);    // tmp = beta * lij
-        TADD(liND, liND, tmpND);       // li = alpha * li + beta * lij (= li_new)
-
-        // Phase 3: Store scalar results to GM (ND format)
-        // mi_new -> mi accumulator, li_new -> li accumulator
-        // alpha -> mij buffer (reuse), beta -> lij buffer (reuse)
-        set_flag(PIPE_V, PIPE_MTE3, EVENT_ID0);
-        wait_flag(PIPE_V, PIPE_MTE3, EVENT_ID0);
-        TSTORE(miGlobalND, miNewND);   // persist mi_new
-        TSTORE(liGlobalND, liND);      // persist li_new
-        TSTORE(mijGlobalND, alphaND);  // temp: alpha to mij buffer
-        TSTORE(lijGlobalND, betaND);   // temp: beta to lij buffer
-
-        // Phase 4: Reload alpha, beta (and li if last) as ColMajor DN
-        set_flag(PIPE_MTE3, PIPE_MTE2, EVENT_ID0);
-        wait_flag(PIPE_MTE3, PIPE_MTE2, EVENT_ID0);
-        TLOAD(alphaDN, mijGlobalDN);  // alpha from mij buffer as DN
-        TLOAD(betaDN, lijGlobalDN);   // beta from lij buffer as DN
-        if (is_last) {
-            TLOAD(liDN, liGlobalDN);  // li_new from li buffer as DN
-        }
-        set_flag(PIPE_MTE2, PIPE_V, EVENT_ID1);
-        wait_flag(PIPE_MTE2, PIPE_V, EVENT_ID1);
-
-        // Phase 5: Scale data tiles using row-broadcast multiply
-        TROWEXPANDMUL(oiTile, oiTile, alphaDN);       // oi *= alpha
-        TROWEXPANDMUL(oiNewTile, oiNewTile, betaDN);  // oi_new *= beta
-        TADD(oiTile, oiTile, oiNewTile);              // oi = alpha*oi + beta*oi_new
-
-        if (is_last) {
-            // Phase 6: Normalize and output
-            TROWEXPANDDIV(oiTile, oiTile, liDN);  // dst = oi / li_new
-            set_flag(PIPE_V, PIPE_MTE3, EVENT_ID1);
-            wait_flag(PIPE_V, PIPE_MTE3, EVENT_ID1);
-            TSTORE(dstGlobal, oiTile);
-        } else {
-            // Phase 6: Store updated accumulators
-            set_flag(PIPE_V, PIPE_MTE3, EVENT_ID1);
-            wait_flag(PIPE_V, PIPE_MTE3, EVENT_ID1);
-            TSTORE(oiGlobal, oiTile);
-        }
-    }
-
-    set_flag(PIPE_MTE3, PIPE_S, EVENT_ID7);
-    wait_flag(PIPE_MTE3, PIPE_S, EVENT_ID7);
-}
-
-extern "C" __aicore__ void kernel_entry(__gm__ int64_t *args) {
-    __gm__ uint8_t *mij = reinterpret_cast<__gm__ uint8_t *>(args[0]);
-    __gm__ uint8_t *lij = reinterpret_cast<__gm__ uint8_t *>(args[1]);
-    __gm__ uint8_t *oi_new = reinterpret_cast<__gm__ uint8_t *>(args[2]);
-    __gm__ uint8_t *mi = reinterpret_cast<__gm__ uint8_t *>(args[3]);
-    __gm__ uint8_t *li = reinterpret_cast<__gm__ uint8_t *>(args[4]);
-    __gm__ uint8_t *oi = reinterpret_cast<__gm__ uint8_t *>(args[5]);
-    int is_first = static_cast<int>(args[6]);
-    int is_last = static_cast<int>(args[7]);
-    __gm__ uint8_t *dst = reinterpret_cast<__gm__ uint8_t *>(args[8]);
-
-    online_update_impl(mij, lij, oi_new, mi, li, oi, is_first, is_last, dst);
-}
