+++ tests/st/a5/host_build_graph/paged_attention/kernels/aic/aic_qk_matmul.cpp
@@ -99,9 +99,12 @@ extern "C" __aicore__ void kernel_entry(__gm__ int64_t *args) {
     __gm__ uint8_t *kj = reinterpret_cast<__gm__ uint8_t *>(args[1]);
     __gm__ uint8_t *sij = reinterpret_cast<__gm__ uint8_t *>(args[2]);
     int q_tile_size = static_cast<int>(args[3]);
-    // args[4] = head_dim (128), args[5] = block_size
+    int head_dim = static_cast<int>(args[4]);
+    // args[5] = block_size
 
-    if (q_tile_size == 16) {
+    if (q_tile_size == 16 && head_dim <= 16) {
+        qk_matmul_impl<16, 16, 16>(qi, kj, sij);
+    } else if (q_tile_size == 16) {
         qk_matmul_impl<16, 128, 128>(qi, kj, sij);
     } else {
         qk_matmul_impl<64, 128, 64>(qi, kj, sij);
