+++ tests/st/a5/host_build_graph/paged_attention/test_paged_attention.py
@@ -94,6 +94,37 @@ class TestPagedAttentionHostBuildGraph(SceneTestCase):
                 "dtype": "bfloat16",
             },
         },
+        {
+            "name": "SmallCase1",
+            "platforms": ["a5sim", "a5"],
+            "config": {"aicpu_thread_num": 3, "block_dim": 3},
+            "params": {
+                "batch": 1,
+                "num_heads": 16,
+                "kv_head_num": 1,
+                "head_dim": 16,
+                "block_size": 16,
+                "context_len": 16,
+                "max_model_len": 256,
+                "dtype": "bfloat16",
+            },
+        },
+        {
+            "name": "SmallCase2",
+            "platforms": ["a5sim", "a5"],
+            "config": {"aicpu_thread_num": 3, "block_dim": 3},
+            "manual": True,
+            "params": {
+                "batch": 1,
+                "num_heads": 16,
+                "kv_head_num": 1,
+                "head_dim": 16,
+                "block_size": 16,
+                "context_len": 64,
+                "max_model_len": 256,
+                "dtype": "bfloat16",
+            },
+        },
     ]
 
     def generate_args(self, params):
