+++ tests/st/a5/tensormap_and_ringbuffer/mixed_example/kernels/aic/kernel_matmul.cpp
