+++ tests/st/a5/tensormap_and_ringbuffer/mixed_example/kernels/aiv/kernel_mul.cpp
