+++ tests/st/a5/tensormap_and_ringbuffer/mixed_example/test_mixed_example.py
