+++ tests/st/a5/tensormap_and_ringbuffer/spmd_basic/kernels/aic/kernel_spmd_read.cpp
