+++ tests/st/a5/tensormap_and_ringbuffer/spmd_basic/kernels/aiv/kernel_spmd_read.cpp
